flask_backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
openai==1.14.2
python-dotenv==1.0.1
google-cloud-firestore
flask-smorest==0.41.0
flask-swagger-ui==3.36.0